requirements.txt
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
google-cloud-logging>=3.10.0
pyyaml>=6.0
